Add files via upload
This is specifically the API function for my Personal Investment App. For security reasons, the corresponding functions that connect to and use PostgreSQL are not included, but effectively the data from this function is parsed if needed and stored. In the future, a general function passing personal information and passwords for PostgreSQL databases (like how the API_key is done in this function) will become arguments so that one, reusable data pipeline is formed.